An esoteric version of Basic
Symbolic is a fork of basic that has replaced basic keywords with a more aesthetically pleasing and intuitive symbol combinations. There had been a previous attempt to recreate a purely symbolic language using davidcallanan-myopl-code tutorial that I called formal and is uploaded on this profile, but it was not deemed to be a good idea. However, after then an Aha moment came and using working MIT licensed code available, the keywords were mainly comprised of 3 symbol combinations : '~', '|' , ':'. FUN was replaced with 'assume' and ended with that. etc... It s a fun and useful language environment to transition from higher level languages to lower level language, and for the symbolic approach, I find is more aesthetically pleasing. Essentially sample.sym and example.myopl are exactly the same srcipt